# Conflicts:
#	USER_GUIDE.md
#	lib/GfxRenderer/GfxRenderer.cpp
#	open-x4-sdk
#	src/activities/reader/EpubReaderActivity.cpp

+++ src/activities/home/FileBrowserActivity.cpp
@@ -16,7 +16,7 @@
 #include "util/BookCacheUtils.h"
 
 namespace {
-constexpr unsigned long GO_HOME_MS = 1000;
+constexpr uint32_t GO_HOME_MS = 1000;
 constexpr size_t NAME_BUFFER_SIZE = 500;
 }  // namespace
